Move methods to more logical places
